requirements.txt
Flask>=2.0.1
nltk>=3.6.2
tensorflow>=2.5.0
numpy>=1.19.5
google-api-core>=1.31.0
google-auth>=1.32.1
google-auth-oauthlib>=0.4.4
google-cloud>=0.34.0
google-cloud-core>=1.7.1
google-cloud-storage>=1.40.0
google-crc32c>=1.1.2
google-pasta>=0.2.0
google-resumable-media>=1.3.1
googleapis-common-protos>=1.53.0
flask-cors>=3.0.10